Factor out the SymbolIndex contract
