+++ build.sbt
@@ -103,6 +103,7 @@ lazy val metaserver = project
     libraryDependencies ++= List(
       "io.github.soc" % "directories" % "5",
       "org.fusesource.leveldbjni" % "leveldbjni-all" % "1.8",
+      "org.xerial" % "sqlite-jdbc" % "3.21.0",
       "io.monix" %% "monix" % "2.3.0",
       "com.lihaoyi" %% "pprint" % "0.5.3",
       "com.thoughtworks.qdox" % "qdox" % "2.0-M7", // for java mtags
